backend/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
lxml
python-multipart
python-jose[cryptography]
passlib[bcrypt]
apscheduler
tinydb
python-dotenv
pdfminer.six
